fibonacci working

+++ execute/ex_jge.hpp
@@ -19,7 +19,7 @@ namespace cai
 
             static constexpr auto flags_v = typename state::flags_t{};
 
-            static constexpr bool should_jmp = ((flags_v.SF ^ flags_v.OF) == 0);
+            static constexpr bool should_jmp = ((static_cast<int>(flags_v.SF) ^ static_cast<int>(flags_v.OF)) == 0);
             using final_regs_state = std::conditional_t<
                     should_jmp,
                     set_reg<typename state::registers_state_t, regs::id_t::EIP, static_cast<uint32_t>(ip)>,
